Patch v1

+++ my-laravel-app/database/migrations/2025_03_27_085931_create_bookings_table.php
@@ -21,9 +21,9 @@ class CreateBookingsTable extends Migration
             $table->dateTime('end_date');
             $table->string(column: 'id');
             $table->string(column: 'branch_code');
-            $table->integer(column: 'patient_id');
+            $table->bigInteger(column: 'patient_id');
             $table->string(column: 'useReward');
-            
+
             $table->text(column: 'remarks');
         });
     }
